Implement NodeGraph to manage component lifecycle and reactive resolution
