+++ agent_sdks/python/a2ui_core/src/a2ui/core/state/surface_model.py
@@ -14,14 +14,12 @@
 
 import copy
 import warnings
-from typing import Any, Dict, Optional, Union
-from ..common.events import EventSource, Subscription, Signal
+from typing import Any, Dict, Optional
+from ..common.events import EventSource
 from .data_model import DataModel
 from .surface_components_model import SurfaceComponentsModel
-from .component_model import ComponentModel
-from .component_node import ComponentNode
 from ..catalog import Catalog
-from ..validating import CatalogSchemaValidator
+from ..catalog.catalog import TComponent, TFunction
 
 
 class SurfaceModel:
@@ -30,33 +28,20 @@ class SurfaceModel:
     def __init__(
         self,
         surface_id: str,
-        catalog: Catalog[Any, Any],
-        theme: Optional[Dict[str, Any]] = None,
+        catalog: Catalog[TComponent, TFunction],
+        theme: Dict[str, Any] = None,
         send_data_model: bool = False,
-        locale: Optional[str] = None,
+        data_model: DataModel = None,
     ):
         self.id = surface_id
         self.catalog = catalog
         self.theme = theme or {}
         self.send_data_model = send_data_model
-        self.locale = locale
 
-        self.data_model = DataModel()
+        self.data_model = data_model or DataModel()
         self.components_model = SurfaceComponentsModel()
         self.on_action = EventSource()
         self.on_error = EventSource()
-
-        # Node Layer attributes
-        self.rootNode: Signal[Optional[ComponentNode]] = Signal(None)
-        self.active_nodes: Dict[str, ComponentNode] = {}
-
-        # Track model changes
-        self._comp_created_sub = self.components_model.on_created.subscribe(
-            self._on_component_created
-        )
-        self._comp_deleted_sub = self.components_model.on_deleted.subscribe(
-            self._on_component_deleted
-        )
 
     def dispatch_action(
         self, payload: Dict[str, Any], source_component_id: str
@@ -88,293 +73,8 @@ class SurfaceModel:
         err_payload["surfaceId"] = self.id
         self.on_error.emit(err_payload)
 
-    def get_or_create_node(self, component_id: str, data_path: str) -> ComponentNode:
-        """Gets or reactively creates a living Node for a component ID at a given data path."""
-        # Calculate unique instance_id
-        if data_path == "/":
-            instance_id = component_id
-        else:
-            norm_path = data_path.rstrip("/") if data_path != "/" else "/"
-            instance_id = f"{component_id}-[{norm_path}]"
-
-        # Return existing cached node
-        if instance_id in self.active_nodes:
-            return self.active_nodes[instance_id]
-
-        component_model = self.components_model.get(component_id)
-        props_signal = Signal({})
-
-        if component_model:
-            node = ComponentNode(
-                instance_id, component_id, component_model.type, data_path, props_signal
-            )
-        else:
-            # Placeholder for progressive rendering
-            node = ComponentNode(
-                instance_id, component_id, "Placeholder", data_path, props_signal
-            )
-
-        self.active_nodes[instance_id] = node
-
-        # If placeholder, it has no properties. It cleans itself from cache on disposal.
-        if not component_model:
-            node.add_cleanup(lambda: self.active_nodes.pop(instance_id, None))
-            return node
-
-        # Set up reactive context and binder
-        from ..rendering.data_context import DataContext
-        from ..rendering.component_context import ComponentContext
-        from ..rendering.generic_binder import GenericBinder
-
-        data_context = DataContext(
-            path=data_path, data_model=self.data_model, catalog=self.catalog
-        )
-        comp_context = ComponentContext(
-            component_model=component_model,
-            data_context=data_context,
-            surface_components=self.components_model,
-            dispatch_action_callback=self.dispatch_action,
-        )
-
-        binder = GenericBinder(comp_context)
-        node._binder = binder
-
-        child_nodes_by_prop = {}
-        template_subs = {}
-
-        def on_properties_changed(resolved_props: Dict[str, Any]) -> None:
-            new_props = {}
-            for k, v in resolved_props.items():
-                new_props[k] = v
-
-            current_resolved = {}
-
-            # Wrap actions into closures
-            for k, v in list(new_props.items()):
-                if k == "action" or (
-                    isinstance(v, dict) and ("event" in v or "functionCall" in v)
-                ):
-                    action_payload = copy.deepcopy(v)
-
-                    # Create closure wrapping resolution and dispatch
-                    def make_action_closure(payload=action_payload):
-                        resolved_payload = data_context.resolve_dynamic_value(payload)
-                        self.dispatch_action(resolved_payload, component_id)
-
-                    new_props[k] = make_action_closure
-
-            # Get reference fields from catalog
-            ref_map = CatalogSchemaValidator.from_catalog(
-                self.catalog
-            ).extract_ref_fields()
-            comp_type = component_model.type if component_model else ""
-            single_refs, list_refs = ref_map.get(comp_type, (set(), set()))
-
-            # Resolve single-child references
-            for single_ref in single_refs:
-                if single_ref in new_props:
-                    child_id = new_props[single_ref]
-                    if isinstance(child_id, str) and child_id:
-                        child_node = self.get_or_create_node(child_id, data_path)
-                        current_resolved[single_ref] = child_node
-                        new_props[single_ref] = child_node
-                    elif child_id is None:
-                        current_resolved[single_ref] = None
-                        new_props[single_ref] = None
-
-            # Resolve list-child references (explicit lists and templated child lists)
-            for list_ref in list_refs:
-                if list_ref in new_props:
-                    val = new_props[list_ref]
-                    if isinstance(val, list):
-                        child_list = []
-                        for item in val:
-                            if isinstance(item, str) and item:
-                                child_list.append(
-                                    self.get_or_create_node(item, data_path)
-                                )
-                            elif isinstance(item, dict) and "child" in item:
-                                resolved_item = copy.deepcopy(item)
-                                item_child_id = item["child"]
-                                if isinstance(item_child_id, str) and item_child_id:
-                                    resolved_item["child"] = self.get_or_create_node(
-                                        item_child_id, data_path
-                                    )
-                                child_list.append(resolved_item)
-                            else:
-                                child_list.append(item)
-                        current_resolved[list_ref] = child_list
-                        new_props[list_ref] = child_list
-
-                    elif (
-                        isinstance(val, dict) and "componentId" in val and "path" in val
-                    ):
-                        template_comp_id = val["componentId"]
-                        template_path = data_context.resolve_path(val["path"])
-
-                        if list_ref in template_subs:
-                            template_subs[list_ref].unsubscribe()
-                            del template_subs[list_ref]
-
-                        spawned_nodes_signal = Signal([])
-                        new_props[list_ref] = spawned_nodes_signal
-
-                        def on_array_changed(array_data: Any) -> None:
-                            old_spawned = child_nodes_by_prop.get(list_ref, [])
-                            if isinstance(old_spawned, list):
-                                for old_node in old_spawned:
-                                    if isinstance(old_node, ComponentNode):
-                                        old_node.dispose()
-
-                            if not isinstance(array_data, list):
-                                child_nodes_by_prop[list_ref] = []
-                                spawned_nodes_signal.value = []
-                                return
-
-                            new_spawned = []
-                            for i in range(len(array_data)):
-                                scoped_path = f"{template_path}/{i}"
-                                node_inst = self.get_or_create_node(
-                                    template_comp_id, scoped_path
-                                )
-                                new_spawned.append(node_inst)
-
-                            child_nodes_by_prop[list_ref] = new_spawned
-                            spawned_nodes_signal.value = new_spawned
-
-                        sub = self.data_model.subscribe(template_path, on_array_changed)
-                        template_subs[list_ref] = sub
-                        on_array_changed(sub.value)
-                        current_resolved[list_ref] = spawned_nodes_signal
-
-            # Compare current_resolved with child_nodes_by_prop to dispose of no-longer-referenced nodes
-            def collect_nodes(value):
-                nodes = set()
-                if isinstance(value, ComponentNode):
-                    nodes.add(value)
-                elif isinstance(value, list):
-                    for item in value:
-                        nodes.update(collect_nodes(item))
-                elif isinstance(value, dict):
-                    for v in value.values():
-                        nodes.update(collect_nodes(v))
-                elif isinstance(value, Signal):
-                    nodes.update(collect_nodes(value.value))
-                return nodes
-
-            old_referenced_nodes = collect_nodes(list(child_nodes_by_prop.values()))
-            new_referenced_nodes = collect_nodes(list(current_resolved.values()))
-
-            removed_nodes = old_referenced_nodes - new_referenced_nodes
-            for removed_node in removed_nodes:
-                removed_node.dispose()
-
-            # Update child_nodes_by_prop
-            for k, v in current_resolved.items():
-                child_nodes_by_prop[k] = v
-
-            node.props.value = new_props
-
-        # Subscribe node to binder updates
-        binder_sub = binder.subscribe(on_properties_changed)
-
-        def cleanup_node():
-            binder_sub.unsubscribe()
-            binder.dispose()
-            for sub in list(template_subs.values()):
-                sub.unsubscribe()
-            template_subs.clear()
-
-            for item in list(child_nodes_by_prop.values()):
-                if isinstance(item, ComponentNode):
-                    item.dispose()
-                elif isinstance(item, list):
-                    for node_in_list in item:
-                        if isinstance(node_in_list, ComponentNode):
-                            node_in_list.dispose()
-                        elif isinstance(node_in_list, dict) and isinstance(
-                            node_in_list.get("child"), ComponentNode
-                        ):
-                            node_in_list["child"].dispose()
-            child_nodes_by_prop.clear()
-            self.active_nodes.pop(instance_id, None)
-
-        node.add_cleanup(cleanup_node)
-        return node
-
-    def _on_component_created(self, component: ComponentModel) -> None:
-        component_id = component.id
-
-        # 1. Identify nodes that need recreation due to a placeholder or type upgrade
-        nodes_to_recreate = []
-        for node in list(self.active_nodes.values()):
-            if node.component_id == component_id:
-                if node.type == "Placeholder" or node.type != component.type:
-                    nodes_to_recreate.append(node)
-
-        # 2. Recreate identified nodes
-        for old_node in nodes_to_recreate:
-            data_path = old_node.data_path
-            old_node.dispose()
-            self.get_or_create_node(component_id, data_path)
-
-        # 3. Update rootNode if root component was created
-        if component_id == "root" and not self.rootNode.value:
-            self.rootNode.value = self.get_or_create_node("root", "/")
-
-        # 4. Rebuild references on any parents referencing this component
-        for active_node in list(self.active_nodes.values()):
-            if active_node.component_id == component_id:
-                continue
-            if hasattr(active_node, "_binder") and active_node._binder:
-                raw_props = active_node._binder.context.component_model.properties
-                if self._references_component(raw_props, component_id):
-                    active_node._binder._rebuild_all_bindings()
-
-    def _on_component_deleted(self, component_id: str) -> None:
-        # 1. Dispose all active nodes for the component
-        nodes_to_delete = [
-            n for n in self.active_nodes.values() if n.component_id == component_id
-        ]
-        for node in nodes_to_delete:
-            node.dispose()
-
-        # 2. Nullify rootNode if root was deleted
-        if component_id == "root":
-            self.rootNode.value = None
-
-        # 3. Rebuild references on other parents referencing this component
-        for active_node in list(self.active_nodes.values()):
-            if hasattr(active_node, "_binder") and active_node._binder:
-                raw_props = active_node._binder.context.component_model.properties
-                if self._references_component(raw_props, component_id):
-                    active_node._binder._rebuild_all_bindings()
-
-    def _references_component(self, raw_props: Any, target_id: str) -> bool:
-        if raw_props == target_id:
-            return True
-        if isinstance(raw_props, dict):
-            return any(
-                self._references_component(v, target_id) for v in raw_props.values()
-            )
-        if isinstance(raw_props, list):
-            return any(
-                self._references_component(item, target_id) for item in raw_props
-            )
-        return False
-
     def dispose(self) -> None:
-        """Disposes of the surface and its resources, cleaning up all active nodes."""
-        if hasattr(self, "_comp_created_sub"):
-            self._comp_created_sub.unsubscribe()
-        if hasattr(self, "_comp_deleted_sub"):
-            self._comp_deleted_sub.unsubscribe()
-
-        for node in list(self.active_nodes.values()):
-            node.dispose()
-        self.active_nodes.clear()
-        self.rootNode.value = None
-
+        """Disposes of the surface and its resources."""
         if hasattr(self.data_model, "dispose") and callable(self.data_model.dispose):
             try:
                 self.data_model.dispose()
